Add files via upload
All User HTML files have the standard bar at the top that will be used as a nav bar to main pages. Not all links are added yet but this is what I will work off of. ProfileU and SearchU already have the CSS done for the nav bar that can be found in profileUStyle.css and SearchStyle.css